move settings button out of sidebar
this fixes the problem where the button would scroll weirdly when the
sidebar was collapsed. I think it looks nicer too.

part of this commit also fixes the HoverTooltip component so that it
can work with fixed sized elements. this works by not creating a
wrapper div around the children. Instead we use the child directly